refactor: reduce Codacy complexity warnings
Component extraction (LogDetailModal):
- Extract StatusBadge to LogDetailStatusBadge.tsx
- Extract DetailItem to LogDetailItem.tsx (LucideIcon type)
- Extract RequestLogDetail to RequestLogDetail.tsx
- LogDetailModal reduced from 577 to ~90 LOC (complexity 28 -> ~5)

Test file splits (by describe block, project convention):
- useArenaState.test.ts (1232 NCL) -> 5 focused test files + helpers
- useConversationRunner.test.ts (732 NCL) -> 7 focused test files + helpers
- ChatMessageList.test.tsx (563 NCL) -> 9 focused test files

CI fixes:
- Use CODACY_API_TOKEN env var for account token auth
- Extract POSTGRES_PASSWORD to job-level env

Config:
- Exclude README.md from markdownlint in .codacy.yml
