--- web/src/pages/Arena/__tests__/useArenaState.test.ts
@@ -1,1589 +1,0 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { act, renderHook } from "@testing-library/react";
-import React from "react";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { useArenaState } from "../useArenaState";
-
-const createQueryClient = () =>
-	new QueryClient({
-		defaultOptions: {
-			queries: { retry: false },
-			mutations: { retry: false },
-		},
-	});
-
-const createWrapper = () => {
-	const queryClient = createQueryClient();
-	return function Wrapper({ children }: { children: React.ReactNode }) {
-		return React.createElement(
-			QueryClientProvider,
-			{ client: queryClient },
-			children,
-		);
-	};
-};
-
-// Mock the dependencies
-vi.mock("../../../hooks/useModels", () => ({
-	useEnabledModels: vi.fn(() => ({
-		data: [
-			{ provider_name: "TestProvider", model_id: "model-1", enabled: true },
-			{ provider_name: "TestProvider", model_id: "model-2", enabled: true },
-			{ provider_name: "TestProvider", model_id: "model-3", enabled: true },
-		],
-	})),
-}));
-
-// vi.hoisted ensures the variable is available when vi.mock factory runs
-// We use a mutable ref so tests can change the arenaSubMode per-test
-const arenaModeRef = vi.hoisted(() => ({
-	current: "compare" as "compare" | "competition",
-}));
-
-// Mutable ref for persistArena - allows tests to toggle persistence
-const persistRef = vi.hoisted(() => ({ current: false }));
-
-// Mutable refs for arenaHistory mocking
-const arenaHistoryMocks = vi.hoisted(() => ({
-	saveCompareToHistory: vi.fn(),
-	getArenaHistoryEnabled: vi.fn(() => false),
-}));
-
-vi.mock("../../../context/SidebarModeContext", () => ({
-	useSidebarMode: vi.fn(() => ({
-		get arenaSubMode() {
-			return arenaModeRef.current;
-		},
-		setArenaSubMode: vi.fn((v: "compare" | "competition") => {
-			arenaModeRef.current = v;
-		}),
-		chatSubMode: "chat" as const,
-		setChatSubMode: vi.fn(),
-		logsSubMode: "request" as const,
-		setLogsSubMode: vi.fn(),
-	})),
-}));
-
-vi.mock("../../../context/ToastContext", () => ({
-	useToast: vi.fn(() => ({
-		toast: vi.fn(),
-	})),
-}));
-
-vi.mock("../../../context/StorageContext", () => ({
-	useStorage: vi.fn(() => ({
-		persistArena: persistRef.current,
-	})),
-}));
-
-vi.mock("../../../utils/arenaHistory", () => ({
-	getArenaHistoryEnabled: () => arenaHistoryMocks.getArenaHistoryEnabled(),
-	saveCompareToHistory: (...args: unknown[]) =>
-		arenaHistoryMocks.saveCompareToHistory(...args),
-}));
-
-describe("useArenaState", () => {
-	beforeEach(() => {
-		localStorage.clear();
-		arenaModeRef.current = "compare";
-	});
-
-	it("initializes with default values", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		expect(result.current.compareModels).toEqual([]);
-		expect(result.current.bracketModels).toEqual([]);
-		expect(result.current.rounds).toEqual([]);
-		expect(result.current.currentRound).toBe(0);
-		expect(result.current.phase).toBe("setup");
-		expect(result.current.runningModels).toEqual(new Set());
-		expect(result.current.winnerModal).toBe(null);
-		expect(result.current.disabledModels).toEqual(new Set());
-		expect(result.current.arenaCollapsed).toBe(false);
-		expect(result.current.pendingFullReset).toBe(false);
-		expect(result.current.showHistoryModal).toBe(false);
-		expect(result.current.modelParams).toEqual({});
-		expect(result.current.paramEditorModel).toBe(null);
-		// arenaMode depends on SidebarModeContext mock - just verify it's a valid mode
-		expect(["compare", "competition"]).toContain(result.current.arenaMode);
-	});
-
-	it("initializes refs with empty values", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		expect(result.current.abortMapRef.current).toEqual(new Map());
-		expect(result.current.lastExtractLenRef.current).toEqual(new Map());
-		expect(result.current.currentRoundRef.current).toBe(0);
-		expect(result.current.roundsLengthRef.current).toBe(0);
-		expect(result.current.roundsRef.current).toEqual([]);
-		expect(result.current.activePromptIdRef.current).toBe(null);
-		expect(result.current.comparePersonaIdRef.current).toBe(null);
-		// arenaMode depends on SidebarModeContext mock - just verify it's a valid mode
-		expect(["compare", "competition"]).toContain(result.current.arenaMode);
-	});
-
-	it("provides setter functions for all state values", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		expect(typeof result.current.setCompareModels).toBe("function");
-		expect(typeof result.current.setBracketModels).toBe("function");
-		expect(typeof result.current.setCompetitionActivePromptId).toBe("function");
-		expect(typeof result.current.setCompareActivePromptId).toBe("function");
-		expect(typeof result.current.setCompetitionPrompt).toBe("function");
-		expect(typeof result.current.setComparePrompt).toBe("function");
-		expect(typeof result.current.setPrompt).toBe("function");
-		expect(typeof result.current.setActivePromptId).toBe("function");
-		expect(typeof result.current.setSavedPrompt).toBe("function");
-		expect(typeof result.current.setComparePersonaId).toBe("function");
-		expect(typeof result.current.setComparePersonaPrompt).toBe("function");
-		expect(typeof result.current.setRounds).toBe("function");
-		expect(typeof result.current.setCurrentRound).toBe("function");
-		expect(typeof result.current.setPhase).toBe("function");
-		expect(typeof result.current.setRunningModels).toBe("function");
-		expect(typeof result.current.setWinnerModal).toBe("function");
-		expect(typeof result.current.setDisabledModels).toBe("function");
-		expect(typeof result.current.setArenaCollapsed).toBe("function");
-		expect(typeof result.current.setPendingFullReset).toBe("function");
-		expect(typeof result.current.setShowHistoryModal).toBe("function");
-		expect(typeof result.current.setModelParams).toBe("function");
-		expect(typeof result.current.setParamEditorModel).toBe("function");
-		expect(typeof result.current.setArenaMode).toBe("function");
-	});
-
-	it("updates compareModels via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setCompareModels(["model-1", "model-2"]);
-		});
-
-		expect(result.current.compareModels).toEqual(["model-1", "model-2"]);
-	});
-
-	it("updates bracketModels via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setBracketModels([
-				"model-1",
-				"model-2",
-				"model-3",
-				"model-4",
-			]);
-		});
-
-		expect(result.current.bracketModels).toEqual([
-			"model-1",
-			"model-2",
-			"model-3",
-			"model-4",
-		]);
-	});
-
-	it("updates competitionActivePromptId via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setCompetitionActivePromptId("prompt-123");
-		});
-
-		expect(result.current.competitionActivePromptId).toBe("prompt-123");
-
-		act(() => {
-			result.current.setCompetitionActivePromptId(null);
-		});
-
-		expect(result.current.competitionActivePromptId).toBe(null);
-	});
-
-	it("updates compareActivePromptId via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setCompareActivePromptId("prompt-456");
-		});
-
-		expect(result.current.compareActivePromptId).toBe("prompt-456");
-
-		act(() => {
-			result.current.setCompareActivePromptId(null);
-		});
-
-		expect(result.current.compareActivePromptId).toBe(null);
-	});
-
-	it("updates competitionPrompt via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setCompetitionPrompt("Competition prompt");
-		});
-
-		expect(result.current.competitionPrompt).toBe("Competition prompt");
-	});
-
-	it("updates comparePrompt via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setComparePrompt("Compare prompt");
-		});
-
-		expect(result.current.comparePrompt).toBe("Compare prompt");
-	});
-
-	it("setPrompt dispatches to active mode", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const isCompareMode = result.current.arenaMode === "compare";
-
-		act(() => {
-			result.current.setPrompt("Test prompt");
-		});
-
-		// Should update the prompt for the active mode
-		if (isCompareMode) {
-			expect(result.current.comparePrompt).toBe("Test prompt");
-		} else {
-			expect(result.current.competitionPrompt).toBe("Test prompt");
-		}
-	});
-
-	it("setActivePromptId dispatches to active mode", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const isCompareMode = result.current.arenaMode === "compare";
-
-		act(() => {
-			result.current.setActivePromptId("prompt-id");
-		});
-
-		// Should update the active prompt ID for the active mode
-		if (isCompareMode) {
-			expect(result.current.compareActivePromptId).toBe("prompt-id");
-		} else {
-			expect(result.current.competitionActivePromptId).toBe("prompt-id");
-		}
-	});
-
-	it("updates savedPrompt via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setSavedPrompt("Saved prompt");
-		});
-
-		expect(result.current.savedPrompt).toBe("Saved prompt");
-	});
-
-	it("updates comparePersonaId via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setComparePersonaId("persona-123");
-		});
-
-		expect(result.current.comparePersonaId).toBe("persona-123");
-
-		act(() => {
-			result.current.setComparePersonaId(null);
-		});
-
-		expect(result.current.comparePersonaId).toBe(null);
-	});
-
-	it("updates comparePersonaPrompt via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setComparePersonaPrompt("Persona system prompt");
-		});
-
-		expect(result.current.comparePersonaPrompt).toBe("Persona system prompt");
-	});
-
-	it("updates rounds via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const newRounds = [
-			{
-				matchups: [
-					{
-						slotA: {
-							modelId: "model-1",
-							personaId: null,
-							personaPrompt: "",
-							params: {},
-						},
-						slotB: null,
-						responseA: null,
-						responseB: null,
-						vote: null,
-					},
-				],
-			},
-		];
-
-		act(() => {
-			result.current.setRounds(newRounds);
-		});
-
-		expect(result.current.rounds).toEqual(newRounds);
-	});
-
-	it("updates currentRound via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setCurrentRound(2);
-		});
-
-		expect(result.current.currentRound).toBe(2);
-	});
-
-	it("updates phase via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setPhase("running");
-		});
-
-		expect(result.current.phase).toBe("running");
-
-		act(() => {
-			result.current.setPhase("voting");
-		});
-
-		expect(result.current.phase).toBe("voting");
-	});
-
-	it("updates runningModels via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setRunningModels(new Set(["model-1", "model-2"]));
-		});
-
-		expect(result.current.runningModels).toEqual(
-			new Set(["model-1", "model-2"]),
-		);
-	});
-
-	it("updates winnerModal via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const winnerModal = {
-			winner: "model-1",
-			rounds: [],
-		};
-
-		act(() => {
-			result.current.setWinnerModal(winnerModal);
-		});
-
-		expect(result.current.winnerModal).toEqual(winnerModal);
-
-		act(() => {
-			result.current.setWinnerModal(null);
-		});
-
-		expect(result.current.winnerModal).toBe(null);
-	});
-
-	it("updates disabledModels via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setDisabledModels(new Set(["model-1"]));
-		});
-
-		expect(result.current.disabledModels).toEqual(new Set(["model-1"]));
-	});
-
-	it("updates arenaCollapsed via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setArenaCollapsed(true);
-		});
-
-		expect(result.current.arenaCollapsed).toBe(true);
-	});
-
-	it("updates pendingFullReset via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setPendingFullReset(true);
-		});
-
-		expect(result.current.pendingFullReset).toBe(true);
-	});
-
-	it("updates showHistoryModal via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setShowHistoryModal(true);
-		});
-
-		expect(result.current.showHistoryModal).toBe(true);
-	});
-
-	it("updates modelParams via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setModelParams({
-				"model-1": { temperature: 0.7, max_tokens: 100 },
-			});
-		});
-
-		expect(result.current.modelParams).toEqual({
-			"model-1": { temperature: 0.7, max_tokens: 100 },
-		});
-	});
-
-	it("updates paramEditorModel via setter", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setParamEditorModel("model-1");
-		});
-
-		expect(result.current.paramEditorModel).toBe("model-1");
-
-		act(() => {
-			result.current.setParamEditorModel(null);
-		});
-
-		expect(result.current.paramEditorModel).toBe(null);
-	});
-
-	it("can update refs directly", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.abortMapRef.current.set("model-1", new AbortController());
-		});
-
-		expect(result.current.abortMapRef.current.size).toBe(1);
-
-		act(() => {
-			result.current.lastExtractLenRef.current.set("key-1", 100);
-		});
-
-		expect(result.current.lastExtractLenRef.current.get("key-1")).toBe(100);
-
-		act(() => {
-			result.current.currentRoundRef.current = 3;
-		});
-
-		expect(result.current.currentRoundRef.current).toBe(3);
-
-		act(() => {
-			result.current.roundsLengthRef.current = 5;
-		});
-
-		expect(result.current.roundsLengthRef.current).toBe(5);
-
-		act(() => {
-			result.current.activePromptIdRef.current = "prompt-ref";
-		});
-
-		expect(result.current.activePromptIdRef.current).toBe("prompt-ref");
-
-		act(() => {
-			result.current.comparePersonaIdRef.current = "persona-ref";
-		});
-
-		expect(result.current.comparePersonaIdRef.current).toBe("persona-ref");
-	});
-
-	it("computes canRun correctly", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const isCompareMode = result.current.arenaMode === "compare";
-
-		// Initially false - no models selected
-		expect(result.current.canRun).toBe(false);
-
-		if (isCompareMode) {
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			// Should be true with 2+ models and prompt
-			expect(result.current.canRun).toBe(true);
-
-			act(() => {
-				result.current.setPrompt("");
-			});
-
-			// False without prompt
-			expect(result.current.canRun).toBe(false);
-		} else {
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			// In competition mode with 2 models (valid bracket size)
-			expect(result.current.canRun).toBe(true);
-		}
-	});
-
-	it("computes disabledReason correctly", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const isCompareMode = result.current.arenaMode === "compare";
-
-		if (isCompareMode) {
-			// Should show reason when no models selected
-			expect(result.current.disabledReason).toContain("Select");
-
-			act(() => {
-				result.current.setCompareModels(["model-1"]);
-			});
-
-			// Should show reason when only 1 model
-			expect(result.current.disabledReason).toContain("more");
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-			});
-
-			// Should show reason when no prompt
-			expect(result.current.disabledReason).toContain("prompt");
-
-			act(() => {
-				result.current.setPrompt("Test prompt");
-			});
-
-			// Should be empty when ready
-			expect(result.current.disabledReason).toBe("");
-		} else {
-			// Competition mode checks bracketModels
-			expect(result.current.disabledReason).toContain("Select");
-		}
-	});
-
-	it("provides buildCompareRoundWithParams function", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const rounds = result.current.buildCompareRoundWithParams(
-			["model-1", "model-2"],
-			null,
-			"",
-		);
-
-		expect(rounds).toHaveLength(1);
-		expect(rounds[0].matchups).toHaveLength(2);
-	});
-
-	it("provides buildInitialRoundsWithParams function", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const rounds = result.current.buildInitialRoundsWithParams([
-			"model-1",
-			"model-2",
-			"model-3",
-			"model-4",
-		]);
-
-		expect(rounds.length).toBeGreaterThan(0);
-	});
-
-	it("handleRandomComparePersona selects a random persona", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.handleRandomComparePersona();
-		});
-
-		// Should have set a persona ID and prompt
-		expect(result.current.comparePersonaId).toBeDefined();
-		expect(result.current.comparePersonaPrompt).toBeDefined();
-	});
-
-	it("handleRandomBracketModel adds a random model", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const initialCount = result.current.bracketModels.length;
-
-		act(() => {
-			result.current.handleRandomBracketModel();
-		});
-
-		// Should have added a model if available
-		expect(result.current.bracketModels.length).toBeGreaterThanOrEqual(
-			initialCount,
-		);
-	});
-
-	it("handleRandomCompareModel adds a random model", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const initialCount = result.current.compareModels.length;
-
-		act(() => {
-			result.current.handleRandomCompareModel();
-		});
-
-		// Should have added a model if available
-		expect(result.current.compareModels.length).toBeGreaterThanOrEqual(
-			initialCount,
-		);
-	});
-
-	it("previewPairs returns pairs in competition mode setup", () => {
-		// This test needs competition mode - set it explicitly BEFORE renderHook
-		arenaModeRef.current = "competition";
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		// Verify we're in competition mode
-		expect(result.current.arenaMode).toBe("competition");
-		// previewPairs is only non-null in competition mode during setup
-		expect(result.current.previewPairs).toBeDefined();
-		// Note: mode will be reset to "compare" by beforeEach of next test
-	});
-
-	it("cleans up abort controllers on unmount", () => {
-		const abortCtrl = new AbortController();
-		const { unmount } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		// Simulate adding abort controllers (this would normally happen during use)
-		// The cleanup happens on unmount
-		unmount();
-
-		// After unmount, abort controllers should be aborted
-		expect(abortCtrl.signal.aborted).toBe(false); // Not added to the hook's map
-	});
-
-	it("syncs roundsRef with rounds state", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		const newRounds = [
-			{
-				matchups: [
-					{
-						slotA: {
-							modelId: "model-1",
-							personaId: null,
-							personaPrompt: "",
-							params: {},
-						},
-						slotB: null,
-						responseA: null,
-						responseB: null,
-						vote: null,
-					},
-				],
-			},
-		];
-
-		act(() => {
-			result.current.setRounds(newRounds);
-		});
-
-		expect(result.current.roundsRef.current).toEqual(newRounds);
-	});
-
-	it("syncs activePromptIdRef with activePromptId state", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setActivePromptId("sync-test-id");
-		});
-
-		expect(result.current.activePromptIdRef.current).toBe("sync-test-id");
-	});
-
-	it("syncs comparePersonaIdRef with comparePersonaId state", () => {
-		const { result } = renderHook(() => useArenaState(), {
-			wrapper: createWrapper(),
-		});
-
-		act(() => {
-			result.current.setComparePersonaId("sync-persona-id");
-		});
-
-		expect(result.current.comparePersonaIdRef.current).toBe("sync-persona-id");
-	});
-
-	describe("canRun - compare mode", () => {
-		beforeEach(() => {
-			arenaModeRef.current = "compare";
-		});
-
-		it("is false with no models", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is false with only 1 model", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is false with duplicate models", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-1"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is true with 2 distinct models and prompt", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(true);
-		});
-
-		it("is false when prompt is empty", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-				result.current.setPrompt("");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is false when prompt is whitespace-only", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-				result.current.setPrompt("   ");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-	});
-
-	describe("disabledReason - compare mode", () => {
-		beforeEach(() => {
-			arenaModeRef.current = "compare";
-		});
-
-		it('shows "Select at least 2 models" when 0 models', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.disabledReason).toBe("Select at least 2 models");
-		});
-
-		it('shows "Pick at least 1 more model" when 1 model', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1"]);
-			});
-
-			expect(result.current.disabledReason).toBe("Pick at least 1 more model");
-		});
-
-		it('shows "No duplicate models" when duplicates present', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-1"]);
-			});
-
-			expect(result.current.disabledReason).toBe("No duplicate models");
-		});
-
-		it('shows "Enter a prompt" when models OK but no prompt', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-			});
-
-			expect(result.current.disabledReason).toBe("Enter a prompt");
-		});
-
-		it("is empty string when all conditions met", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setCompareModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.disabledReason).toBe("");
-		});
-	});
-
-	describe("canRun - competition mode", () => {
-		beforeEach(() => {
-			arenaModeRef.current = "competition";
-		});
-
-		it("is false with 0 models", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is false with 1 model", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is false with 3 models (invalid bracket size)", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2", "model-3"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is false with 5 models (invalid bracket size)", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels([
-					"model-1",
-					"model-2",
-					"model-3",
-					"model-4",
-					"model-5",
-				]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-
-		it("is true with 2 models + prompt", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(true);
-		});
-
-		it("is true with 4 models + prompt", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels([
-					"model-1",
-					"model-2",
-					"model-3",
-					"model-4",
-				]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(true);
-		});
-
-		it("is true with 8 models + prompt", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels([
-					"model-1",
-					"model-2",
-					"model-3",
-					"model-4",
-					"model-5",
-					"model-6",
-					"model-7",
-					"model-8",
-				]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(true);
-		});
-
-		it("is false with duplicate models", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-1"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.canRun).toBe(false);
-		});
-	});
-
-	describe("disabledReason - competition mode", () => {
-		beforeEach(() => {
-			arenaModeRef.current = "competition";
-		});
-
-		it('shows "Select 2, 4, or 8 models" with 0 models', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.disabledReason).toBe("Select 2, 4, or 8 models");
-		});
-
-		it('shows "Pick at least 1 more model" with 1 model', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1"]);
-			});
-
-			expect(result.current.disabledReason).toBe("Pick at least 1 more model");
-		});
-
-		it('shows "No duplicate models" with duplicates', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-1"]);
-			});
-
-			expect(result.current.disabledReason).toBe("No duplicate models");
-		});
-
-		it("shows pick-or-remove message for 3 models", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2", "model-3"]);
-			});
-
-			// nextBracketSize(3) = 4, so "Pick 1 more or remove to get 4"
-			expect(result.current.disabledReason).toBe(
-				"Pick 1 more or remove to get 4",
-			);
-		});
-
-		it('shows "Enter a prompt" when models OK but no prompt', () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2"]);
-			});
-
-			expect(result.current.disabledReason).toBe("Enter a prompt");
-		});
-
-		it("is empty string when all conditions met", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-			});
-
-			expect(result.current.disabledReason).toBe("");
-		});
-
-		it("shows voting phase message", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2"]);
-				result.current.setPrompt("Test prompt");
-				result.current.setPhase("voting");
-			});
-
-			expect(result.current.disabledReason).toBe(
-				"Vote on all matchups to continue to the next round",
-			);
-		});
-
-		it("shows next round prompt message when next_round_ready with no prompt", () => {
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setBracketModels(["model-1", "model-2"]);
-				result.current.setPrompt("");
-				result.current.setPhase("next_round_ready");
-			});
-
-			expect(result.current.disabledReason).toBe(
-				"Enter a prompt for the next round",
-			);
-		});
-	});
-
-	describe("localStorage initialization with persistArena=true", () => {
-		beforeEach(() => {
-			localStorage.clear();
-			arenaModeRef.current = "compare";
-			// Enable persistence
-			persistRef.current = true;
-		});
-
-		afterEach(() => {
-			// Reset to default
-			persistRef.current = false;
-		});
-
-		it("initializes compareModels from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					compareModels: ["P1/M1", "P2/M2"],
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.compareModels).toEqual(["P1/M1", "P2/M2"]);
-		});
-
-		it("initializes bracketModels from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					bracketModels: ["P1/M1", "P2/M2", "P3/M3", "P4/M4"],
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.bracketModels).toEqual([
-				"P1/M1",
-				"P2/M2",
-				"P3/M3",
-				"P4/M4",
-			]);
-		});
-
-		it("initializes bracketModels from legacy group1Models/group2Groups fallback", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					group1Models: ["P1/M1", "P2/M2"],
-					group2Models: ["P3/M3", "P4/M4"],
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.bracketModels).toEqual([
-				"P1/M1",
-				"P2/M2",
-				"P3/M3",
-				"P4/M4",
-			]);
-		});
-
-		it("initializes savedPrompt from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					savedPrompt: "My saved prompt",
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.savedPrompt).toBe("My saved prompt");
-		});
-
-		it("initializes rounds from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			const mockRounds: import("../types").BracketRound[] = [
-				{
-					matchups: [
-						{
-							slotA: {
-								modelId: "model-1",
-								personaId: null,
-								personaPrompt: "",
-								params: {},
-							},
-							slotB: null,
-							responseA: null,
-							responseB: null,
-							vote: null,
-						},
-					],
-				},
-			];
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					rounds: mockRounds,
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.rounds).toEqual(mockRounds);
-		});
-
-		it("initializes currentRound from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					currentRound: 2,
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.currentRound).toBe(2);
-		});
-
-		it("initializes phase from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					phase: "running",
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.phase).toBe("running");
-		});
-
-		it("initializes arenaCollapsed from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					arenaCollapsed: true,
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.arenaCollapsed).toBe(true);
-		});
-
-		it("initializes modelParams from localStorage", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					modelParams: {
-						"model-1": { temperature: 0.8, max_tokens: 200 },
-					},
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.modelParams).toEqual({
-				"model-1": { temperature: 0.8, max_tokens: 200 },
-			});
-		});
-
-		it("falls back to defaults when localStorage parse fails", () => {
-			localStorage.setItem("persistArena", "true");
-			localStorage.setItem("arenaState", "invalid json");
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			expect(result.current.compareModels).toEqual([]);
-			expect(result.current.bracketModels).toEqual([]);
-			expect(result.current.savedPrompt).toBe("");
-			expect(result.current.rounds).toEqual([]);
-			expect(result.current.currentRound).toBe(0);
-			expect(result.current.phase).toBe("setup");
-			expect(result.current.arenaCollapsed).toBe(false);
-			expect(result.current.modelParams).toEqual({});
-		});
-	});
-
-	describe("localStorage initialization with persistArena=false", () => {
-		beforeEach(() => {
-			localStorage.clear();
-			arenaModeRef.current = "compare";
-			// Ensure persistence is disabled
-			persistRef.current = false;
-		});
-
-		it("falls back to defaults when persistArena=false", () => {
-			// Don't set persistArena in localStorage - leave it null
-			// The initializers check localStorage.getItem("persistArena") === "true"
-			// so without it set, they should return defaults
-			localStorage.setItem(
-				"arenaState",
-				JSON.stringify({
-					compareModels: ["P1/M1"],
-					bracketModels: ["P1/M1"],
-					savedPrompt: "test",
-					phase: "running",
-				}),
-			);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			// Should use defaults since persistArena is not "true" in localStorage
-			expect(result.current.compareModels).toEqual([]);
-			expect(result.current.bracketModels).toEqual([]);
-			expect(result.current.savedPrompt).toBe("");
-			expect(result.current.phase).toBe("setup");
-		});
-	});
-
-	describe("compare history save effect", () => {
-		beforeEach(() => {
-			localStorage.clear();
-			arenaModeRef.current = "compare";
-			// Reset to default
-			persistRef.current = false;
-			// Reset mocks
-			arenaHistoryMocks.saveCompareToHistory.mockClear();
-			arenaHistoryMocks.getArenaHistoryEnabled.mockReturnValue(false);
-		});
-
-		it("saves compare history when phase becomes finished in compare mode", async () => {
-			arenaHistoryMocks.getArenaHistoryEnabled.mockReturnValue(true);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			// Set up a round with responses - must include all required ArenaResponse fields
-			const mockRounds: import("../types").BracketRound[] = [
-				{
-					matchups: [
-						{
-							slotA: {
-								modelId: "model-1",
-								personaId: null,
-								personaPrompt: "",
-								params: {},
-							},
-							slotB: null,
-							responseA: {
-								done: true,
-								model: "model-1",
-								rawContent: "Raw Response A",
-								content: "Response A",
-								thinkingContent: "Thinking A",
-								startTimeMs: 1000,
-								error: null,
-								metrics: {
-									tokensPerSecond: 10,
-									durationMs: 1000,
-									promptTokens: 50,
-									completionTokens: 100,
-								},
-							},
-							responseB: null,
-							vote: null,
-						},
-					],
-				},
-			];
-
-			// First set rounds - this will sync roundsRef
-			act(() => {
-				result.current.setRounds(mockRounds);
-			});
-
-			// Flush the roundsRef sync effect before changing phase
-			await act(async () => {});
-
-			// Now set phase to finished - this triggers the effect
-			act(() => {
-				result.current.setPhase("finished");
-			});
-
-			// Flush pending effects
-			await act(async () => {});
-
-			expect(arenaHistoryMocks.saveCompareToHistory).toHaveBeenCalled();
-			expect(arenaHistoryMocks.getArenaHistoryEnabled).toHaveBeenCalled();
-		});
-
-		it("does NOT save compare history when arenaMode is competition", async () => {
-			arenaHistoryMocks.getArenaHistoryEnabled.mockReturnValue(true);
-			arenaModeRef.current = "competition";
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setRounds([{ matchups: [] }]);
-				result.current.setPhase("finished");
-			});
-
-			await act(async () => {});
-
-			expect(arenaHistoryMocks.saveCompareToHistory).not.toHaveBeenCalled();
-		});
-
-		it("does NOT save compare history when getArenaHistoryEnabled is false", async () => {
-			arenaHistoryMocks.getArenaHistoryEnabled.mockReturnValue(false);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			act(() => {
-				result.current.setRounds([{ matchups: [] }]);
-				result.current.setPhase("finished");
-			});
-
-			await act(async () => {});
-
-			expect(arenaHistoryMocks.saveCompareToHistory).not.toHaveBeenCalled();
-		});
-
-		it("resets compareHistorySavedRef when phase leaves finished", async () => {
-			arenaHistoryMocks.getArenaHistoryEnabled.mockReturnValue(true);
-
-			const { result } = renderHook(() => useArenaState(), {
-				wrapper: createWrapper(),
-			});
-
-			// Set up minimal rounds
-			const mockRounds: import("../types").BracketRound[] = [
-				{
-					matchups: [
-						{
-							slotA: {
-								modelId: "model-1",
-								personaId: null,
-								personaPrompt: "",
-								params: {},
-							},
-							slotB: null,
-							responseA: {
-								done: true,
-								model: "model-1",
-								rawContent: "Response",
-								content: "Response",
-								thinkingContent: "",
-								startTimeMs: 1000,
-								error: null,
-								metrics: {
-									tokensPerSecond: 10,
-									durationMs: 1000,
-									promptTokens: 50,
-									completionTokens: 100,
-								},
-							},
-							responseB: null,
-							vote: null,
-						},
-					],
-				},
-			];
-
-			// First set rounds
-			act(() => {
-				result.current.setRounds(mockRounds);
-			});
-
-			// Flush the roundsRef sync effect
-			await act(async () => {});
-
-			// Set phase to finished - should trigger save
-			act(() => {
-				result.current.setPhase("finished");
-			});
-			await act(async () => {});
-
-			expect(arenaHistoryMocks.saveCompareToHistory).toHaveBeenCalledTimes(1);
-
-			// Go back to setup - should reset the flag
-			act(() => {
-				result.current.setPhase("setup");
-			});
-			await act(async () => {});
-
-			// Now if we go to finished again, it should save again (flag was reset)
-			act(() => {
-				result.current.setPhase("finished");
-			});
-			await act(async () => {});
-
-			// Should have been called twice now
-			expect(arenaHistoryMocks.saveCompareToHistory).toHaveBeenCalledTimes(2);
-		});
-	});
-});
